requirements.txt
# Core Framework Dependencies
python-nmap==1.5.1
masscan==0.1.6
pwntools==4.11.1
requests==2.31.0
aiohttp==3.9.1
asyncio-throttle==1.0.2

# Binary Analysis & Exploit Development
angr==9.2.77
r2pipe==1.7.4
capstone==5.0.1
keystone-engine==0.9.2
unicorn==2.0.1.post1

# Database & Storage
sqlalchemy==2.0.23
redis==5.0.1
pymongo==4.6.0

# Vulnerability Intelligence
nvdlib==0.7.4
shodan==1.30.1
censys==2.2.8

# Network & Protocol Libraries
scapy==2.5.0
paramiko==3.4.0
impacket==0.11.0
dnspython==2.4.2
cryptography==41.0.8

# Web & API
flask==3.0.0
flask-restful==0.3.10
beautifulsoup4==4.12.2
selenium==4.16.0

# Utilities & Logging
colorama==0.4.6
rich==13.7.0
click==8.1.7
pyyaml==6.0.1
python-dotenv==1.0.0
loguru==0.7.2

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0